# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ChannelRef": {
        "additionalProperties": false,
        "description": "The channel a job's video came from; jobs with one are queued for posting (ADR-23).",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "title": "ChannelRef",
        "type": "object"
      },
      "ClipOptions": {
        "additionalProperties": false,
        "properties": {
          "caption_style": {
            "const": "default",
            "default": "default",
            "title": "Caption Style",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "max_len": {
            "default": 60.0,
            "maximum": 180.0,
            "title": "Max Len",
            "type": "number"
          },
          "min_len": {
            "default": 30.0,
            "minimum": 5.0,
            "title": "Min Len",
            "type": "number"
          },
          "min_score": {
            "default": 0.8,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "type": "number"
          },
          "n": {
            "anyOf": [
              {
                "maximum": 30.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N"
          },
          "reframe": {
            "default": "auto",
            "enum": [
              "auto",
              "center",
              "blur"
            ],
            "title": "Reframe",
            "type": "string"
          }
        },
        "title": "ClipOptions",
        "type": "object"
      },
...
      "CostSummary": {
        "properties": {
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageCost"
            },
            "title": "Stages",
            "type": "array"
          },
          "total_usd": {
            "readOnly": true,
            "title": "Total Usd",
            "type": "number"
          }
        },
        "required": [
          "total_usd"
        ],
        "title": "CostSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobError": {
        "properties": {
          "error_type": {
            "title": "Error Type",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/StageName"
          }
        },
        "required": [
          "stage",
          "error_type",
          "message"
        ],
        "title": "JobError",
        "type": "object"
      },
      "JobInput": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChannelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "notify": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelegramTarget"
              },
              {
                "type": "null"
              }
            ]
          },
          "options": {
            "$ref": "#/components/schemas/ClipOptions"
          },
          "permission": {
            "$ref": "#/components/schemas/Permission"
          },
          "source_credit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Credit"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "telegram_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram File Id"
          }
        },
        "required": [
          "permission"
        ],
        "title": "JobInput",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "done",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobView": {
        "description": "What the API returns: the core record merged with its clips, and the total cost.",
        "properties": {
          "clips": {
            "items": {
              "$ref": "#/components/schemas/ClipState"
            },
            "title": "Clips",
            "type": "array"
          },
          "cost": {
            "$ref": "#/components/schemas/CostSummary"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "output_zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Zip"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Progress"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageName"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "stage",
          "clips",
          "cost",
          "created_at",
          "updated_at"
        ],
        "title": "JobView",
        "type": "object"
      },
      "Permission": {
        "description": "Permission basis for using a source (docs/SOURCING.md).",
        "enum": [
          "own",
          "creator_agreement",
          "clipping_program",
          "cc_by",
          "public_domain"
        ],
        "title": "Permission",
        "type": "string"
      },
...
      "PostingOverview": {
        "additionalProperties": false,
        "description": "What `/status`, `GET /posting` and `clipforge status` show (spec \u00a79).",
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountPosting"
            },
            "title": "Accounts",
            "type": "array"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelProgress"
            },
            "title": "Channels",
            "type": "array"
          },
          "days_left": {
            "title": "Days Left",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "next_slot": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Slot"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "per_day": {
            "title": "Per Day",
            "type": "integer"
          },
          "problem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "waiting": {
            "title": "Waiting",
            "type": "integer"
          }
        },
        "required": [
          "enabled",
          "paused",
          "channels",
          "waiting",
          "days_left",
          "per_day",
          "next_slot"
        ],
        "title": "PostingOverview",
        "type": "object"
      },
...
      "StageName": {
        "enum": [
          "ingest",
          "transcribe",
          "highlights",
          "reframe",
          "captions",
          "render",
          "package"
        ],
        "title": "StageName",
        "type": "string"
      },
      "TelegramTarget": {
        "additionalProperties": false,
        "description": "Where the Telegram notifier reports a job (the chat that sent it).",
        "properties": {
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "reply_to_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To Message Id"
          }
        },
        "required": [
          "chat_id"
        ],
        "title": "TelegramTarget",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ClipForge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/jobs": {
      "post": {
        "operationId": "post_job_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Post Job Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Job"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/jobs/{job_id}/download": {
      "get": {
        "operationId": "download_jobs__job_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exp",
            "required": true,
            "schema": {
              "title": "Exp",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": true,
            "schema": {
              "title": "Sig",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download"
      }
    },
    "/jobs/{job_id}/resume": {
      "post": {
        "operationId": "post_resume_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Resume"
      }
    },
    "/posting": {
      "get": {
        "operationId": "get_posting_posting_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Posting"
      }
    },
    "/posting/rebuild": {
      "post": {
        "operationId": "post_posting_rebuild_posting_rebuild_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Post Posting Rebuild Posting Rebuild Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Posting Rebuild"
      }
    },
    "/posting/restore": {
      "post": {
        "operationId": "post_posting_restore_posting_restore_post",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Post Posting Restore Posting Restore Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Posting Restore"
      }
    },
    "/telegram/webhook": {
      "post": {
        "operationId": "telegram_webhook_telegram_webhook_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-telegram-bot-api-secret-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Bot-Api-Secret-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Telegram Webhook Telegram Webhook Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telegram Webhook"
      }
    }
  }
}